site/requirements.txt
asgiref==3.4.1
bcrypt==3.2.0
cffi==1.14.6
Django==3.2.7
django-cors-headers==3.8.0
mysqlclient==2.0.3
Pillow==8.3.2
pycparser==2.20
PyJWT==2.2.0
python-dotenv==0.19.0
pytz==2021.1
six==1.16.0
sqlparse==0.4.2
